+++ service/controller/clusterapi/controller.go
@@ -8,22 +8,20 @@ import (
 	"github.com/giantswarm/operatorkit/v7/pkg/resource"
 	promclient "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned"
 	vpa_clientset "k8s.io/autoscaler/vertical-pod-autoscaler/pkg/client/clientset/versioned"
-	capiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/domain"
 	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/project"
 	controllerresource "github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource"
 )
 
 type ControllerConfig struct {
-	K8sClient        k8sclient.Interface
-	Logger           micrologger.Logger
-	PrometheusClient promclient.Interface
-	VpaClient        vpa_clientset.Interface
-
-	HTTPProxy  string
-	HTTPSProxy string
-	NoProxy    string
+	K8sClient          k8sclient.Interface
+	Logger             micrologger.Logger
+	PrometheusClient   promclient.Interface
+	VpaClient          vpa_clientset.Interface
+	ProxyConfiguration domain.ProxyConfiguration
 
 	AdditionalScrapeConfigs string
 	Bastions                []string
@@ -77,7 +75,7 @@ func NewController(config ControllerConfig) (*Controller, error) {
 			Logger:    config.Logger,
 			Name:      project.Name() + "-cluster-api-controller",
 			NewRuntimeObjectFunc: func() client.Object {
-				return new(capiv1alpha3.Cluster)
+				return new(capi.Cluster)
 			},
 			Resources: resources,
 		}