+++ CHANGELOG.md
@@ -7,9 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.5.1] - 2022-08-24
+
+### Fixed
+
+- Fix CAPI MCs being seen as workload cluster.
+
+## [4.5.0] - 2022-08-24
+
+### Changed
+
+- Change CAPI version from v1alpha3 to v1beta1.
+
+## [4.4.1] - 2022-08-19
+
+### Fixed
+
+- Fix Team hydra config.
+
+## [4.4.0] - 2022-08-17
+
 ### Added
 
 - Add service priority as a tag in opsgenie alerts.
+- Add Team Hydra receiver and route.
 
 ### Fixed
 
@@ -46,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgrade operatorkit from v7.0.1 to v7.1.0.
+- Upgrade github.com/sirupsen/logrus from 1.8.1 to 1.9.0.
 
 ### Added
 
@@ -1562,7 +1584,11 @@ This release was created on release-v3.5.x branch to fix release 3.6.0 see PR#99
 - First release.
 
 
-[Unreleased]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.3.0...HEAD
+[Unreleased]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.5.1...HEAD
+[4.5.1]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.5.0...v4.5.1
+[4.5.0]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.4.1...v4.5.0
+[4.4.1]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.4.0...v4.4.1
+[4.4.0]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.3.0...v4.4.0
 [4.3.0]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.2.0...v4.3.0
 [4.2.0]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.1.0...v4.2.0
 [4.1.0]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.0.1...v4.1.0